:sparkles: (flake.nix) Update flake

+++ flake.nix
@@ -195,18 +195,18 @@
             LD_LIBRARY_PATH = "${pkgs.lib.makeLibraryPath [ pkgs.stdenv.cc.cc ]}";
 
             shellHook = ''
-              rm -r ./fonts
-              mkdir -p ./fonts/noto-fonts
+              	      [ -e ./fonts ] && rm -r ./fonts
+                            mkdir -p ./fonts/noto-fonts
 
-              # unlink .textlintrc
-              ln -s ${textlintrc} .textlintrc
+                            [ -f .textlintrc ] && rm .textlintrc
+                            ln -s ${textlintrc} .textlintrc
 
-              ln -s ${fonts}/bin/NotoSansCJKjp-Bold.otf ./fonts/noto-fonts/NotoSansCJKjp-Bold.otf
-              ln -s ${fonts}/bin/NotoSansCJKjp-Bold.otf ./fonts/noto-fonts/NotoSansCJKjp-Black.otf
-              ln -s ${fonts}/bin/NotoSansCJKjp-Bold.otf ./fonts/noto-fonts/NotoSansCJKjp-Regular.otf
+                            ln -s ${fonts}/bin/NotoSansCJKjp-Bold.otf ./fonts/noto-fonts/NotoSansCJKjp-Bold.otf
+                            ln -s ${fonts}/bin/NotoSansCJKjp-Bold.otf ./fonts/noto-fonts/NotoSansCJKjp-Black.otf
+                            ln -s ${fonts}/bin/NotoSansCJKjp-Bold.otf ./fonts/noto-fonts/NotoSansCJKjp-Regular.otf
 
 
-              ${pkgs.git-secrets}/bin/git-secrets --add '''^[a-z]{4}-[a-z]{4}-[a-z]{4}-[a-z0-9]{4}$'
+                            ${pkgs.git-secrets}/bin/git-secrets --add '''^[a-z]{4}-[a-z]{4}-[a-z]{4}-[a-z0-9]{4}$'
             '';
           };
         };
